change vatRate to BigDecimal for precise calclation

+++ src/main/java/org/example/entity/invoice/CreateInvoiceDTO.java
@@ -1,5 +1,6 @@
 package org.example.entity.invoice;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.UUID;
@@ -9,6 +10,6 @@ public record CreateInvoiceDTO(
     UUID clientId,
     String number,
     LocalDateTime dueDate,
-    Float vatRate,
+    BigDecimal vatRate,
     List<InvoiceItemDTO> items
 ) {}
